+++ src/collective/contact/plonegroup/safe_utils.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+
+from collective.contact.plonegroup.utils import get_organization
+from collective.contact.plonegroup.utils import get_organizations
+from collective.contact.plonegroup.utils import get_own_organization
+from collective.contact.plonegroup.utils import get_plone_group
+from collective.contact.plonegroup.utils import get_plone_group_id
+from collective.contact.plonegroup.utils import organizations_with_suffixes